Rename & move pane files
