+++ tasks-app-shared/src/androidMain/kotlin/net/opatry/tasks/app/ui/screen/taskListScreenAndroid.kt
@@ -38,8 +38,6 @@ import net.opatry.tasks.app.ui.TaskListsViewModel
 import net.opatry.tasks.app.ui.component.LoadingPane
 import net.opatry.tasks.app.ui.component.NoTaskListEmptyState
 import net.opatry.tasks.app.ui.component.NoTaskListSelectedEmptyState
-import net.opatry.tasks.app.ui.component.TaskListDetail
-import net.opatry.tasks.app.ui.component.TaskListsColumn
 import net.opatry.tasks.resources.Res
 import net.opatry.tasks.resources.default_task_list_title
 import org.jetbrains.compose.resources.stringResource
