+++ tasks-app-shared/src/commonMain/kotlin/net/opatry/tasks/app/ui/screen/tasksPane.kt
@@ -20,7 +20,7 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.opatry.tasks.app.ui.component
+package net.opatry.tasks.app.ui.screen
 
 import CalendarDays
 import CheckCheck
@@ -116,6 +116,14 @@ import kotlinx.datetime.plus
 import kotlinx.datetime.toLocalDateTime
 import kotlinx.datetime.todayIn
 import net.opatry.tasks.app.ui.TaskListsViewModel
+import net.opatry.tasks.app.ui.component.EditTextDialog
+import net.opatry.tasks.app.ui.component.EmptyState
+import net.opatry.tasks.app.ui.component.MissingScreen
+import net.opatry.tasks.app.ui.component.RowWithIcon
+import net.opatry.tasks.app.ui.component.TaskListMenu
+import net.opatry.tasks.app.ui.component.TaskListMenuAction
+import net.opatry.tasks.app.ui.component.TaskMenu
+import net.opatry.tasks.app.ui.component.TaskMenuAction
 import net.opatry.tasks.app.ui.model.DateRange
 import net.opatry.tasks.app.ui.model.TaskListUIModel
 import net.opatry.tasks.app.ui.model.TaskUIModel
