+++ tasks-app-shared/src/jvmMain/kotlin/net/opatry/tasks/app/ui/screen/taskListScreenJvm.kt
@@ -37,8 +37,6 @@ import net.opatry.tasks.app.ui.TaskListsViewModel
 import net.opatry.tasks.app.ui.component.LoadingPane
 import net.opatry.tasks.app.ui.component.NoTaskListEmptyState
 import net.opatry.tasks.app.ui.component.NoTaskListSelectedEmptyState
-import net.opatry.tasks.app.ui.component.TaskListDetail
-import net.opatry.tasks.app.ui.component.TaskListsColumn
 import net.opatry.tasks.resources.Res
 import net.opatry.tasks.resources.default_task_list_title
 import org.jetbrains.compose.resources.stringResource
